// src/app/exports/enums.ts
export enum PAWLMainRoutes {
     CLIENT = '', // Default route
     ADMIN = 'panel'
}

export enum PAWLClientRoutes {
     HOME = 'home',
     MY_PETS = 'my-pets',
     SHOP = 'shop',
     PROFILE = 'profile',
     ADD_PET = 'add-pet',
     PAYMENT = 'payment'
}

export enum PAWLLocalStorage {
     SELECTED_PET_ID = 'selectedPetID',
     SELECTED_PET_TYPE_ID = 'selectedPetTypeID',
     SELECTED_CATEGORY_TYPE_ID = 'selectedCategoryID',
     SELECTED_SUBCATEGORY_ID = 'selectedSubCategoryID',
     SELECTED_PRODUCT_ID = 'selectedProductID',
     CART = 'cart',
     SELECTED_PET = 'selectedPet',
}